Cleaned up WS and related issues, and - added:   comments, 'toc/' and 'images/' skips, description,   total file check count, and a few error checks - removed:   dot printout and 'figures/' skip

+++ gen-phpweb-sqlite-db.php
@@ -1,38 +1,41 @@
-<?php # $Id$
-
+<?php
 /*
+ $Id$
 
-   This is a CLI script; call it thus:
-   
+This script scans files within the generated PHP Manual (phpweb theme)
+and creates shortcut lookups for them with their metaphone based
+companions. Mirrors with SQLite then check against this data. So,
+this allows the likes of "php.net/foo" to work.
+
+This script is known to be imperfect, and future plans will use PhD to
+generate this data based on XML ids thus including ini settings,
+predefined constants, functions, etc.
+
+Usage:
 	php gen-phpweb-sqlite-db.php database DOCUMENT_ROOT MANUAL_PREFIX [implied_lang]
 
-   eg:
-   	php gen-phpweb-sqlite-db.php database /local/Web/sites/phpweb /manual
+Example:
+	php gen-phpweb-sqlite-db.php database /local/Web/sites/phpweb /manual
 
+Each week this information is rendered on the rsync box.
 */
 
-$DOCUMENT_ROOT = $argv[2];
-$DOCUMENT_ROOT_LEN = strlen($DOCUMENT_ROOT);
-$MANUAL_PREFIX = $argv[3];
-$implied_lang = isset($argv[4]) ? $argv[4] : false;
+$DOCUMENT_ROOT		= $argv[2];
+$DOCUMENT_ROOT_LEN	= strlen($DOCUMENT_ROOT);
+$MANUAL_PREFIX		= $argv[3];
+$implied_lang		= isset($argv[4]) ? $argv[4] : false;
+$db_name			= $argv[1];
 
-$db_name = $argv[1];
+$s = sqlite_open($db_name);
+if (!$s) {
+	return;
+}
 
 if (!file_exists($db_name)) {
-	$s = sqlite_open($db_name);
-
-	sqlite_query($s, "CREATE TABLE fs (
-lang char(5) not null,
-prefix char(32) not null,
-keyword char(128) not null,
-name varchar(238) not null,
-prio int not null
-);");
-	sqlite_query($s, "CREATE INDEX map ON fs (lang,keyword)");
-	sqlite_query($s, "BEGIN");
+	if (!create_url_database($s)) {
+		return;
+	}
 } else {
-	$s = sqlite_open($db_name);
-	/* purge old results */
 	sqlite_query($s, "BEGIN");
 	sqlite_query($s, "delete from fs");
 }
@@ -44,31 +47,34 @@ if ($implied_lang !== false) {
 }
 
 sqlite_query($s, "COMMIT");
-
 exit(0);
-
 
 function scan($dir, $lang)
 {
 	global $s;
-	static $sections = array('book.', 'ref.', 'function.', 'class.', 'feature-',
-		'control-structures.', 'language.',
-		"about.", "faq.", "features.", 
-	);
 	global $DOCUMENT_ROOT_LEN;
+	static $sections = array(
+		'book.', 'ref.', 'function.', 'class.', 'feature-',
+		'control-structures.', 'language.',
+		'about.', 'faq.', 'features.',
+	);
 	
-	$dot = 0;
+	$count = 0;
 	echo "Lang: $lang\n";
 
 	$d = opendir($dir);
 	if (!$d) {
 		return;
 	}
-	readdir($d); readdir($d); // skip . and ..
+
 	while (($f = readdir($d)) !== false) {
-		if ($f === 'CVS' || $f === 'figures') {
+
+		// Directories to skip
+		$skips = array('CVS', 'images', 'toc', '.', '..');
+		if (in_array($f, $skips)) {
 			continue;
 		}
+
 		$file = $dir . DIRECTORY_SEPARATOR . $f;
 
 		if (is_dir($file)) {
@@ -77,32 +83,44 @@ function scan($dir, $lang)
 			/* which section/prefix does this fall under ? */
 			$prefix = "";
 			$keyword = $f;
-		
+
+			// Get filepath relative to the manual/
 			$doc_rel = substr($file, $DOCUMENT_ROOT_LEN);
-		
+
+			// Keyword is filename minus the extension
 			$x = strrpos($keyword, '.');
 			if ($x !== false) {
 				$keyword = substr($keyword, 0, $x);
 			}
-			
+
+			/* Example:
+			- section:	book.
+			- keyword:	about.formats
+			- f:		about.formats.php
+			- doc_rel:	/en/about.formats.php
+			- x: 13
+			*/
 			foreach ($sections as $prio => $section) {
+				/* Example:
+				- before: keyword: about.formats prefix:        section: about. f: about.formats.php
+				- after:  keyword: formats       prefix: about. section: about. f: about.formats.php
+				*/
 				if (!strncmp($f, $section, strlen($section))) {
 					$keyword = substr($keyword, strlen($section));
 					$prefix = $section;
 					break;
 				}
 			}
-			if (++$dot % 76 == 0) {
-				$dot = 0;
-				echo "\n";
-			}
-			echo ".";
-//			echo "$lang: $keyword\n";
+			++$count;
+
 			sqlite_query($s, "INSERT INTO fs (lang, prefix, keyword, name, prio) values ('$lang', '$prefix', '$keyword', '$doc_rel', $prio)");
 			sqlite_query($s, "INSERT INTO fs (lang, prefix, keyword, name, prio) values ('$lang', '$prefix', '" . metaphone($keyword) . "', '$doc_rel', ".($prio+10).")");
+
 		}
 	}
 	closedir($d);
+	
+	echo "Added entries for $count files\n";
 	echo "\n";
 }
 
@@ -128,5 +146,17 @@ function scan_langs($root)
 	closedir($d);
 }
 
-
-?>
+function create_url_database ($s) {
+	$q = "
+		CREATE TABLE fs (
+			lang char(5) not null,
+			prefix char(32) not null,
+			keyword char(128) not null,
+			name varchar(238) not null,
+			prio int not null
+		);
+		";
+	sqlite_query($s, $q);
+	sqlite_query($s, "CREATE INDEX map ON fs (lang,keyword)");
+	sqlite_query($s, "BEGIN");
+}
